+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 3.1.1 under development
 
-- no changes in this release.
+- New #249 Add to the 'Router' middleware option to ignore method failure handle (@olegbaturin)
+- New #249 Add to the 'Router' middleware custom response factories for the method failure responses (@olegbaturin)
 
 ## 3.1.0 February 20, 2024